london_crime_backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
pandas>=2.0
openpyxl>=3.1
requests>=2.31
